+++ tests/cli/admin/test_citus_distribution.py
@@ -208,7 +208,7 @@ def test_mixed_distribution(
         default_distribution_strategy="schema",
         model_distribution={
             "distribute/data/New": {
-                "distribute": "copy",
+                "distribute": ["copy"],
             },
             "distribute/data/Data": {
                 "distribute": {
@@ -302,7 +302,7 @@ def test_default_schema_distribution_invalidation(
     assert not {new_table_identifier}.issubset(updated_citus_state.local)
 
     result = cli.invoke(
-        rc.fork({"models": {"distribute/data/Data": {"distribute": "copy"}}}),
+        rc.fork({"models": {"distribute/data/Data": {"distribute": ["copy"]}}}),
         ["admin", Script.CITUS_DISTRIBUTION.value],
     )
     assert result.exit_code == 0